tests/requirements-test.txt
pytest~=8.3.3
python-dotenv~=1.0.1
pytest-asyncio~=0.24.0
